# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book sale",
        "description": "Простой сервис по пробаже книг",
        "version": "1"
    },
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Получение списка не проданных книг",
                "description": "Получение списка не проданных книг",
                "operationId": "28b9632495ac5f6f66b2e52001d510e6",
                "parameters": [
                    {
                        "parameter": "page",
                        "name": "page",
                        "in": "query",
                        "description": "Текущая страница",
                        "required": false,
                        "deprecated": false,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "paginator": {
                                            "$ref": "#/components/schemas/Paginator"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "text"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "text"
                                                    },
                                                    "cover_url": {
                                                        "type": "string",
                                                        "format": "url",
                                                        "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "1.2"
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "authors": {
                                                        "type": "array",
                                                        "items": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "format": "int",
                                                                    "example": "1"
                                                                },
                                                                "first_name": {
                                                                    "type": "string",
                                                                    "format": "first name",
                                                                    "example": "Oleg"
                                                                },
                                                                "last_name": {
                                                                    "type": "string",
                                                                    "format": "last name",
                                                                    "example": "Oleg"
                                                                },
                                                                "rank": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": "1.24"
                                                                },
                                                                "avatar_url": {
                                                                    "type": "string",
                                                                    "format": "url",
                                                                    "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "books"
                ],
                "summary": "Создание книги",
                "description": "Создание книги",
                "operationId": "b10f5fe35eefa47e974dd4b333d870b1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "text"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "text"
                                    },
                                    "cover_url": {
                                        "type": "string",
                                        "format": "url",
                                        "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "1.2"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "format": "int",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "text"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "format": "string",
                                                        "example": "text"
                                                    },
                                                    "cover_url": {
                                                        "type": "string",
                                                        "format": "url",
                                                        "example": "http://rodriguez.com/aut-similique-laudantium-qui-assumenda-ducimus-quisquam"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "1.2"
                                                    },
                                                    "quantity": {
                                                        "type": "integer",
                                                        "format": "int",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/books/{id}/buy": {
            "post": {
                "tags": [
                    "books"
                ],
                "summary": "Покупка книги",
                "description": "Покупка книги",
                "operationId": "1f85e8c8c6add52e9c4dc42f9f6cfd1f",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "query",
                        "description": "Идентификатор книги",
                        "required": true,
                        "deprecated": false,
                        "allowEmptyValue": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Paginator": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "first_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=1"
                    },
                    "from": {
                        "type": "integer",
                        "format": "int",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "380"
                    },
                    "last_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=5"
                    },
                    "next_page_url": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=3"
                    },
                    "path": {
                        "type": "string",
                        "format": "url",
                        "example": "http://138.124.55.208/api/prod/chat?page=2"
                    },
                    "per_page": {
                        "type": "integer",
                        "format": "int",
                        "example": "30"
                    },
                    "to": {
                        "type": "integer",
                        "format": "int",
                        "example": "30"
                    },
                    "total": {
                        "type": "integer",
                        "format": "int",
                        "example": "11371"
                    }
                },
                "type": "object"
            }
        }
    },
...
}